+++ apps/web/src/entities/feedback/lib/apis/index.ts
@@ -1,0 +1,2 @@
+export { getPartFeedback } from './get-part-feedback'
+export { getPathFeedback } from './get-path-feedback'
