+++ apps/web/src/entities/feedback/lib/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useGetPartFeedback } from './use-get-part-feedback'
+export { useGetPathFeedback } from './use-get-path-feedback'
