+++ apps/web/src/entities/feedback/lib/hooks/use-get-part-feedback.ts
@@ -1,0 +1,5 @@
+export function useGetPartFeedback() {
+  const onGetPartFeedback = () => {}
+
+  return { onGetPartFeedback }
+}
