+++ apps/web/src/entities/feedback/lib/hooks/use-get-path-feedback.ts
@@ -1,0 +1,5 @@
+export function useGetPathFeedback() {
+  const onGetPathFeedback = () => {}
+
+  return { onGetPathFeedback }
+}
