+++ apps/web/src/features/feedback/lib/apis/index.ts
@@ -1,0 +1,2 @@
+export { postPartFeedback } from './post-part-feedback'
+export { postPathFeedback } from './post-path-feedback'
