+++ apps/web/src/features/feedback/lib/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useSubmitPartFeedback } from './use-submit-part-feedback'
+export { useSubmitPathFeedback } from './use-submit-path-feedback'
