+++ apps/web/src/features/feedback/lib/hooks/use-submit-part-feedback.ts
@@ -1,0 +1,5 @@
+export function useSubmitPartFeedback() {
+  const onSubmitFeedback = () => {}
+
+  return { onSubmitFeedback }
+}
