+++ apps/web/src/features/feedback/lib/hooks/use-submit-path-feedback.ts
@@ -1,0 +1,5 @@
+export function useSubmitPathFeedback() {
+  const onSubmitFeedback = () => {}
+
+  return { onSubmitFeedback }
+}
